Moved `computedProp` inside `ts/types/utils.ts`
